+++ hawkbit-repository/hawkbit-repository-api/src/main/java/org/eclipse/hawkbit/repository/event/remote/MultiActionEvent.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2015 Bosch Software Innovations GmbH and others.
+ * Copyright (c) 2019 Bosch Software Innovations GmbH and others.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -8,36 +8,51 @@
  */
 package org.eclipse.hawkbit.repository.event.remote;
 
+import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 /**
- * Deployment event which implies that the Multi-Assignment feature is enabled.
+ * Generic deployment event for the Multi-Assignments feature. The event payload
+ * holds a list of controller IDs identifying the targets which are affected by
+ * a deployment action (e.g. a software assignment (update) or a cancellation of
+ * an update).
  */
-@SuppressWarnings("squid:MaximumInheritanceDepth")
-public class MultiActionEvent extends DeploymentEvent {
+public class MultiActionEvent extends RemoteTenantAwareEvent implements Iterable<String> {
 
     private static final long serialVersionUID = 1L;
 
+    private final List<String> controllerIds = new ArrayList<>();
+
     /**
-     * Default constructor for serialization purposes.
+     * Default constructor.
      */
     public MultiActionEvent() {
-        super();
+        // for serialization libs like jackson
     }
 
     /**
-     * Constructs a Multi-Action event for the targets with the given controller
-     * IDs.
+     * Constructor.
      * 
      * @param tenant
-     *            this event is scoped to
+     *            tenant the event is scoped to
      * @param applicationId
-     *            of the application
+     *            the application id
      * @param controllerIds
-     *            of the targets this event refers to
+     *            the controller IDs of the affected targets
      */
     public MultiActionEvent(final String tenant, final String applicationId, final List<String> controllerIds) {
-        super(tenant, applicationId, controllerIds);
+        super(applicationId, tenant, applicationId);
+        this.controllerIds.addAll(controllerIds);
+    }
+
+    public List<String> getControllerIds() {
+        return controllerIds;
+    }
+
+    @Override
+    public Iterator<String> iterator() {
+        return controllerIds.iterator();
     }
 
 }
